requirements.txt
accelerate==1.4.0
diffusers==0.32.2
einops==0.8.1
huggingface-hub==0.29.1
image-reward==1.5
imageio==2.37.0
matplotlib==3.10.0
natsort==8.4.0
numpy==1.26.3
omegaconf==2.3.0
pillow==11.0.0
scikit-learn==1.6.1
scipy==1.15.2
sentencepiece==0.2.0
supervision==0.25.1
tokenizers==0.21.0
tqdm==4.67.1
transformers==4.49.0